added resource_browser.ui simplified code and remove not required functions added automatic icon scaling when browser is resized
